site_backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic-settings==2.6.0
psycopg[binary]==3.2.1
bcrypt==4.1.3
PyJWT==2.9.0
pytest==8.3.3
